License the flattened mapper as basic.
