+++ docs/reference/mapping/types/flattened.asciidoc
@@ -1,3 +1,6 @@
+[role="xpack"]
+[testenv="basic"]
+
 [[flattened]]
 === Flattened datatype
