+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/flattened/FlattenedFeatureSetUsage.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+package org.elasticsearch.xpack.core.flattened;
+
+import org.elasticsearch.common.io.stream.StreamInput;
+import org.elasticsearch.xpack.core.XPackFeatureSet;
+import org.elasticsearch.xpack.core.XPackField;
+
+import java.io.IOException;
+
+public class FlattenedFeatureSetUsage extends XPackFeatureSet.Usage {
+
+    public FlattenedFeatureSetUsage(StreamInput input) throws IOException {
+        super(input);
+    }
+
+    public FlattenedFeatureSetUsage(boolean available, boolean enabled) {
+        super(XPackField.FLATTENED, available, enabled);
+    }
+}
