+++ x-pack/plugin/src/test/resources/rest-api-spec/test/flattened/10_basic.yml
